python_rag/requirements.txt
langchain==0.3.13
langchain-community==0.3.13
langchain-groq==0.2.8
faiss-cpu==1.9.0
sentence-transformers==3.3.1
pypdf==5.1.0
python-dotenv==1.0.1
torch==2.5.1
transformers==4.47.1